backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
passlib
bcrypt==4.0.1
requests
pydantic
pymongo